Final Submission
Build notes:
-Added notes to all important files I created or modified.
-Changed SQL database name, included new SQL file.
-Changed references in code from old database name to the new name
(menu_entry_db).
-Added a new variable in MenuDay.java, which corresponds with the ID
field of each weekly_menu DB entry. A method has also been made which
converts this ID number to the correct day of the week.  This is used to
display the day on the web page to view the menu.
-Changed text and messages to reflect that this is a cafeteria menu web
app.
-GitHub only says changes will be made to the new and deleted SQL files.
If for some reason none of the other changes listed are there, please
let me know and I will send the build via Google Drive.